==> 2.Differential_ATAC-seq_analysis_all.Rmd <==
---
title: "Differential analysis of atac-seq peaks"
output:
  html_document:
    df_print: paged
---
# Differential atac-seq analysis as with RNA-seq data.

```{r setup, include=FALSE, warning=FALSE}
library(edgeR)
library(knitr)   # for kable tables
library(ggplot2)
library(grid)
library(gridExtra)
library(reshape)
library(plyr)
library(geneplotter)
library(pheatmap)
library(RColorBrewer)
library(org.Hs.eg.db)
library(limma)
library(ggbiplot)
library(ggrepel) # provides geoms for ggplot2 to repel overlapping text labels.

#library(ggfortify)
library(reshape2)  # to modify dataframes for ggplot2
library(knitr)   #for tables in rmd
library(pander)  #more tables

currentDate <- Sys.Date() # to save date in name of output files

sample = factor(c("sbad2.1", "sbad3.1", "neo1.1"),
levels = c("sbad2.1", "sbad3.1", "neo1.1"))   # data comes from three donors

stage = factor(
c("iPSC", "DE", "GT", "PF", "PE", "EP", "EN", "BLC"),
levels = c("iPSC", "DE", "GT", "PF", "PE", "EP", "EN", "BLC")
) # 8 stages


```


After checking that everything looks OK and selecting the right normalization and CPM cutoff, we'll do the differential  analysis. This is going to be a "peak" differential analysis. It highlights peaks with counts that are more different per stage to all other stages. This is different to a contrast that would compare each stage to iPSC, for example. To do the second analysis one needs to change only the design matrix. For more info on how to do contrasts, see scripts for RNA-seq diff expression analysis. 

```{r voom and plots, warning=FALSE}
load(file = "/Users/Marta/Documents/WTCHG/DPhil/Data/Regulation/atac-seq/session_objects/dge_atac-seq_10CPM_trim_conservativeCounts_allQualitySamples.xz", verbose =
       TRUE)  #loading again the dge object to do the selected QC
       
       
       
       samples <- rep(sample, 8)
       
       stages <- rep(stage, each = 3)
       
       #create the design matrix
       design1 <- model.matrix( ~ stages + samples)
       kable(design1)
       
       
       # This converts the counts to log-counts per million with associated precision weights. After this, the RNA-seq data can be analyzed as if it was microarray data.
       
       
       v2 = voom(dge, design = design1, plot = TRUE)
       
       
       # For first 50 peaks, see the distribution of counts:
       random_sample = sample(nrow(dge), 50)
       par(mfrow = c(1, 3))
       plot(density(as.numeric(unlist(dge$counts[random_sample, ]))), main = "counts", cex.main =
       2)
       plot(density(as.numeric(unlist(
       log2(dge$counts + 1)[random_sample, ]
       ))), main = "log2", cex.main = 2)
       plot(density(as.numeric(unlist(v2$E[random_sample, ]))), main = "voom norm counts", cex.main =
       2)
       
       
       #function wrap so that plotMDS doesn't produce a plot when called:
       plotMDS.invisible <- function(...) {
       ff <- tempfile()
       png(filename = ff)
       res <- plotMDS(...)
       dev.off()
       unlink(ff)
       res
       }
       
       pretty_mds = function(v2) {
       mds_p = plotMDS.invisible(v2$E, gene.selection = "pairwise")    # pairwise method (default)
       mds_c = plotMDS.invisible(v2$E, gene.selection = "common")      #common method
       
       
       # Rearrange data for ggplot
       
       # method: pairwise
       m_p = as.data.frame(mds_p$cmdscale.out)
       m_p <- cbind(m_p, sample = samples, stage = stages)
       colnames(m_p) = c("Dimension 1", "Dimension 2", "Samples", "Stages")
       
       # method: common
       m_c = as.data.frame(mds_c$cmdscale.out)
       m_c <- cbind(m_c, sample = samples, stage = stages)
       colnames(m_c) = c("Dimension 1", "Dimension 2", "Samples", "Stages")
       
       
       # plot pairwise
       
       # if(qc_plots==TRUE){
       #tiff("/Users/Marta/Documents/WTCHG/DPhil/Plots/conservative_counts/Diff_v2_MDS_pairwise_only_ours.tiff", type="cairo",
       # width=10,height=10,units="in",res=300,pointsize = 13,compression="lzw")
       mp = ggplot(m_p) +
       geom_point(aes(`Dimension 1` , `Dimension 2`),
       size = 2,
       color = 'grey') +
       geom_label_repel(
       aes(
       `Dimension 1` ,
       `Dimension 2`,
       fill = factor(Samples),
       label = Stages
       ),
       fontface = 'bold',
       color = 'white',
       box.padding = unit(0.25, "lines"),
       point.padding = unit(0.25, "lines")
       ) +
       coord_fixed(ratio = 1.2) +
       
       theme_bw(base_size = 20) +
       theme(
       plot.background = element_blank(),
       panel.grid.major = element_blank(),
       panel.grid.minor = element_blank()
       ) +
       theme(panel.border = element_rect()) +
       
       theme(legend.position = "bottom")  +
       ggtitle("MDS plot:pairwise method") +
       labs(fill = "Sample")
       # dev.off()
       
       # plot common
       # tiff("/Users/Marta/Documents/WTCHG/DPhil/Plots/conservative_counts/Diff_v2_MDS_common_only_ours.tiff", type="cairo",
       #      width=10,height=10,units="in",res=300,pointsize = 13,compression="lzw")
       mc = ggplot(m_c) +
       geom_point(aes(`Dimension 1` , `Dimension 2`),
       size = 2,
       color = 'grey') +
       geom_label_repel(
       aes(
       `Dimension 1` ,
       `Dimension 2`,
       fill = factor(Samples),
       label = Stages
       ),
       fontface = 'bold',
       color = 'white',
       box.padding = unit(0.25, "lines"),
       point.padding = unit(0.25, "lines")
       ) +
       coord_fixed(ratio = 2) +  #fix x-y ratio
       
       theme_bw(base_size = 20) +
       theme(
       plot.background = element_blank(),
       panel.grid.major = element_blank(),
       panel.grid.minor = element_blank()
       ) +
       theme(panel.border = element_rect()) +
       
       theme(legend.position = "bottom")  +
       ggtitle("MDS plot: common method") +
       labs(fill = "Sample")
       # dev.off()
       
       #  }
       return(list(mp, mc))   #return both plots as a list
       }
       
mds_plots = pretty_mds(v2)

plot(mds_plots[[1]])
plot(mds_plots[[2]])


# get sample distance cluster 

plot_sdc = function(x) {
  sampleDists <- dist(t(x))   
  #This function computes and returns the distance matrix computed by using the specified distance measure to compute the distances between the rows of a data matrix.
  # by default, "euclidean"
  sampleDistMatrix <- as.matrix(sampleDists)
  rownames(sampleDistMatrix) <- paste(stages, samples, sep = " - ")
  colnames(sampleDistMatrix) <- NULL
  colors <- colorRampPalette(rev(brewer.pal(9, "Blues")))(255)
  
  png(
  "/Users/Marta/Documents/WTCHG/DPhil/Plots/atac-seq/distance_matrix_atac-seq_10CPM_trim_conservativeCounts_allQualitySamples.png",
  type = "cairo",
  width = 7,
  height = 5,
  units = "in",
  res = 200,
  pointsize = 13
  )
  
  sdc = pheatmap(
  sampleDistMatrix,
  clustering_distance_rows = sampleDists,
  clustering_distance_cols = sampleDists,
  col = colors
  )
  
  dev.off()
  
  return(sdc)
}


sdc_voom = plot_sdc(v2$E)

# spearman rho correlation matrix

spearman_cor = function(x){
  
  cormat <- round(cor(x,method = "spearman"),2)
  reorder_cormat <- function(cormat) {
    # Use correlation between variables as distance
    dd <- as.dist((1 - cormat) / 2)
    hc <- hclust(dd)
    cormat <- cormat[hc$order, hc$order]
  }
  cormat = reorder_cormat(cormat)
  
  melted_cormat <- melt(cormat)
  melted_cormat$Var1 = factor(melted_cormat$Var1,levels = paste(stages,samples,sep = "-"), ordered = T)
  melted_cormat$Var2 = factor(melted_cormat$Var2,levels = paste(stages,samples,sep = "-"), ordered = T)

  
  
  ggheatmap <- ggplot(melted_cormat, aes(Var1, Var2, fill = value)) +
    geom_tile(color = "white") +
    scale_fill_gradientn(
    limit = c(0.40, 1),
     values = scales::rescale(c(0.4, 1.0)),
    colors = c("#faffff", "#002147"), 
    space = "Lab",
    name = "Spearman\nCorrelation"
    ) +
    scale_x_discrete(breaks = c("iPSC-sbad3.1", "DE-sbad3.1","GT-sbad3.1", "PF-sbad3.1", "PE-sbad3.1",  "EP-sbad3.1", "EN-sbad3.1",  "BLC-sbad3.1"), labels = c("iPSC","DE", "GT", "PF", "PE" ,"EP","EN", "BLC")) +
    scale_y_discrete(breaks = c("iPSC-sbad3.1", "DE-sbad3.1","GT-sbad3.1", "PF-sbad3.1", "PE-sbad3.1",  "EP-sbad3.1", "EN-sbad3.1",  "BLC-sbad3.1"), labels = c("iPSC","DE", "GT", "PF", "PE" ,"EP","EN", "BLC")) +
    theme_minimal() + # minimal theme
    theme(
    axis.title = element_blank(),
    axis.text.x = element_text(
    vjust = 1,
    size = 12,
    hjust = 0.5,
    face = "bold"
    ),
    axis.text.y =  element_text(face = "bold", size = 12),
    legend.direction = "horizontal",
    legend.position = "top") +
    coord_fixed()
  
  
  png(
  "/Users/Marta/Documents/WTCHG/DPhil/Plots/atac-seq/Spearman_cor_atac-seq_10CPM_trim_conservativeCounts_allQualitySamples.png",
  type = "cairo",
  width = 7,
  height = 5,
  units = "in",
  res = 200,
  pointsize = 13
  )
  plot(ggheatmap)
  dev.off()
}
  

# pearson correlation matrix


pearson_cor = function(x){
  
 
  cormat <- round(cor(x,method = "pearson"),2)
  # reorder_cormat <- function(cormat) {
  #   # Use correlation between variables as distance
  #   dd <- as.dist((1 - cormat) / 2)
  #   hc <- hclust(dd)
  #   cormat <- cormat[hc$order, hc$order]
  # }
  #cormat = reorder_cormat(cormat)
  
  melted_cormat <- melt(cormat)
  melted_cormat$Var1 = factor(melted_cormat$Var1,levels = paste(stages,samples,sep = "-"), ordered = T)
  melted_cormat$Var2 = factor(melted_cormat$Var2,levels = paste(stages,samples,sep = "-"), ordered = T)

  ggheatmap <- ggplot(melted_cormat, aes(Var1, Var2, fill = value)) +
    geom_tile(color = "white") +
    scale_fill_gradientn(
    limit = c(0.40, 1),
 values = scales::rescale(c(0.4, 1.0)),
    colors = c("#faffff", "#002147"), 
    space = "Lab",
    name = "Pearson\nCorrelation"
    ) +
    scale_x_discrete(breaks = c("iPSC-sbad3.1", "DE-sbad3.1","GT-sbad3.1", "PF-sbad3.1", "PE-sbad3.1",  "EP-sbad3.1", "EN-sbad3.1",  "BLC-sbad3.1"), labels = c("iPSC","DE", "GT", "PF", "PE" ,"EP","EN", "BLC")) +
    scale_y_discrete(breaks = c("iPSC-sbad3.1", "DE-sbad3.1","GT-sbad3.1", "PF-sbad3.1", "PE-sbad3.1",  "EP-sbad3.1", "EN-sbad3.1",  "BLC-sbad3.1"), labels = c("iPSC","DE", "GT", "PF", "PE" ,"EP","EN", "BLC")) +
    theme_minimal() + # minimal theme
    theme(
    axis.title = element_blank(),
    axis.text.x = element_text(
    vjust = 1,
    size = 12,
    hjust = 0.5,
    face = "bold"
    ),
    axis.text.y =  element_text(face = "bold", size = 12),
    legend.direction = "horizontal",
    legend.position = "top") +
    coord_fixed()
  
  png(
  "/Users/Marta/Documents/WTCHG/DPhil/Plots/atac-seq/Pearson_cor_atac-seq_10CPM_trim_conservativeCounts_allQualitySamples.png",
  type = "cairo",
  width = 7,
  height = 5,
  units = "in",
  res = 200,
  pointsize = 13
  )
  
  plot(ggheatmap)
  dev.off()
}
  
spearman_cor(v2$E)
pearson_cor(v2$E)

  
  

par(mfrow = c(1, 1))
plot_pca = function(x, s = samples, st = stages) {
pca1 <- prcomp(t(x), retx = TRUE, scale. = F)

plot(pca1, type = "l") #variance vs first 10 components
summary(pca1)          #importance of each component (important line is "proportion of variance")

percentVar <- (pca1$sdev) ^ 2 / sum(pca1$sdev ^ 2)
percentVar <- round(100 * percentVar)
pcs <- as.data.frame(pca1$x)
pcs <- cbind(pcs, sample = samples, stage = stages)
pcs$stage <- ordered(pcs$stage, levels = stage)

diaPalette <-
    c(
      "#000000",
      "#CADAE8",
      "#7883BA",
      "#755A91",
      "#CC85B1",
      "#F4B8B0",
      "#96665A",
      "#96165A"
    )  # Diabetologia pallete
p <- ggplot(pcs, aes(PC1, PC2, colour = stage, shape = samples)) +
geom_point(size = 4) + xlab(paste0("PC1:" , percentVar[1], "% variance")) +
  scale_color_manual(values = diaPalette, name  ="Stage") +
ylab(paste0("PC2: ", percentVar[2], "% variance")) +
      scale_shape_discrete(name  ="Sample")
p <-
p + theme(
panel.grid.major = element_blank(),
panel.grid.minor = element_blank(),
panel.background = element_blank(),
panel.border = element_rect(fill = NA, colour = "black"),
legend.key = element_blank(),
# legend.position = c(0.5,0.5),
axis.title.y = element_text(
angle = 90,
size = 18,
vjust = 0.2
),
axis.title.x = element_text(
size = 18,
vjust = 0
),
axis.text.x = element_text(

colour = "black",
angle = 90,
size = 16,
vjust = 0.2,
hjust = 1
),

axis.text.y = element_text(colour = "black", size = 16),
legend.text = element_text(colour = "black", size = 16),
legend.title = element_text(colour = "black", size = 16),
axis.ticks = element_line(colour = "black"),
axis.line = element_line(colour = "black")
)
return(p)
}

p = plot_pca(v2$E)
p

ggsave(
paste(
"/Users/Marta/Documents/WTCHG/DPhil/Plots/atac-seq/PCA_atac-seq_10CPM_trim_conservativeCounts_allQualitySamples_",
currentDate,
".jpg",
sep = ""
),
p,
width = 6,
height = 5,
units = "in",
dpi = 350
)


```{r Diff atac, warning=FALSE}
###DEA per se



d = as.data.frame(design1)

stagesiPSC = as.factor(c(1, 1, 1, rep(0, 21)))  # add iPSC to design matrix, because the function to make it ignores this (is the baseline)
d = cbind(d, stagesiPSC)
d = d[c(1, 11, 2:10)] # re-order to loop through it without testing intercept or samples

#initialize list of data frames to save them inside the loop

DE_list <- list()

# loop through stages iPSC to BLC
for (i in colnames(d[2:9])) {
test = as.factor(d[, which(colnames(d) == i)])
design2 = model.matrix( ~ samples + test)
v2 = voom(dge, design = design2, save.plot = TRUE) # voom normalize before fitting the linear model


name <- gsub("stages", "", i)  # name of this stage


fit2 <- lmFit(v2, design2)  # fit the linear model
fit2 <- eBayes(fit2)

diff_exp = topTable(
fit2,
coef = ncol(design2),
sort.by = "none",
number = nrow(fit2$coefficients)
)

#save here table for later

DE_list[[i]] <- diff_exp

diff_exp_sig = diff_exp[which(diff_exp$adj.P.Val < 0.01), ]  #pval<0.01, for plots

# mean-difference plot
plotMD(
fit2,
coef = ncol(design2),
status = rownames(fit2) %in% rownames(diff_exp_sig),
main = paste("Mean-difference plot:", name, "vs other stages"),
ylab = "log-fold change"
)

print("In red, peaks that are significantly (p<0.01) differentially *expressed*")


# Q-Q plot of moderated t-statistics. Points off the line may be differentially expressed

datap = as.data.frame(rownames(fit2))
datap$Colour = "black"
# Set new column values to appropriate colours
datap$Colour[rownames(fit2) %in% rownames(diff_exp_sig)] = "red"

qqt(
fit2$t[, 4],
df = fit2$df.total,
col = datap$Colour,
main = paste("Q-Q plot of moderated t-statistics:", name, "vs other stages")
)
abline(0, 1)

print("In red, peaks that are significantly (p<0.01) differentially *expressed*")

# Volcano plot
volcanoplot(
fit2,
coef = ncol(design2),
highlight = 10,
names = rownames(fit2),
main = paste("Volcano plot:", name, "vs other stages")
)

}

#take logFC and adj P values of each, and combine in single dataframe

combined_df <-
lapply(DE_list, "[", c(1, 5))   # subsetting list of dataframes with columns I want

combined_df <-
do.call("cbind", combined_df)   # merging into one dataframe

#get max conditions per peak
maxVals <-
apply(combined_df, 1, function(x)
which.max(x[c(1, 3, 5, 7, 9, 11, 13, 15)]))

#find significant peaks per stage
#rowsums checks that there's at least one positive logFC
# maxvals selects which one to take as max value for each stage
DE_stages <- list()

sig_iPSC_stage <-
combined_df[combined_df$stagesiPSC.adj.P.Val < 0.01 &
maxVals == 1, c(1, 2)]
sig_iPSC_stage = sig_iPSC_stage[order(sig_iPSC_stage[2]), ] #order by adj p values
sig_iPSC_stage$peak_names = rownames(sig_iPSC_stage)
rownames(sig_iPSC_stage) = NULL # take out row names
sig_iPSC_stage = sig_iPSC_stage[c(3, 1, 2)]
DE_stages[["iPSC"]] <- sig_iPSC_stage

sig_DE_stage <-
combined_df[combined_df$stagesDE.adj.P.Val < 0.01 &
maxVals == 2, c(3, 4)]
sig_DE_stage = sig_DE_stage[order(sig_DE_stage[2]), ]
sig_DE_stage$peak_names = rownames(sig_DE_stage)
rownames(sig_DE_stage) = NULL # take out row names
sig_DE_stage = sig_DE_stage[c(3, 1, 2)]
DE_stages [["DE"]] <- sig_DE_stage

sig_GT_stage <-
combined_df[combined_df$stagesGT.adj.P.Val < 0.01 &
maxVals == 3, c(5, 6)]
sig_GT_stage = sig_GT_stage[order(sig_GT_stage[2]), ]
sig_GT_stage$peak_names = rownames(sig_GT_stage)
rownames(sig_GT_stage) = NULL # take out row names
sig_GT_stage = sig_GT_stage[c(3, 1, 2)]
DE_stages [["GT"]] <- sig_GT_stage

sig_PF_stage <-
combined_df[combined_df$stagesPF.adj.P.Val < 0.01 &
maxVals == 4, c(7, 8)]
sig_PF_stage = sig_PF_stage[order(sig_PF_stage[2]), ]
sig_PF_stage$peak_names = rownames(sig_PF_stage)
rownames(sig_PF_stage) = NULL # take out row names
sig_PF_stage = sig_PF_stage[c(3, 1, 2)]
DE_stages [["PF"]] <- sig_PF_stage

sig_PE_stage <-
combined_df[combined_df$stagesPE.adj.P.Val < 0.01 &
maxVals == 5, c(9, 10)]
sig_PE_stage = sig_PE_stage[order(sig_PE_stage[2]), ]
sig_PE_stage$peak_names = rownames(sig_PE_stage)
rownames(sig_PE_stage) = NULL # take out row names
sig_PE_stage = sig_PE_stage[c(3, 1, 2)]
DE_stages [["PE"]] <- sig_PE_stage

sig_EP_stage <-
combined_df[combined_df$stagesEP.adj.P.Val < 0.01 &
maxVals == 6, c(11, 12)]
sig_EP_stage = sig_EP_stage[order(sig_EP_stage[2]), ]
sig_EP_stage$peak_names = rownames(sig_EP_stage)
rownames(sig_EP_stage) = NULL # take out row names
sig_EP_stage = sig_EP_stage[c(3, 1, 2)]
DE_stages [["EP"]] <- sig_EP_stage

sig_EN_stage <-
combined_df[combined_df$stagesEN.adj.P.Val < 0.01 &
maxVals == 7, c(13, 14)]
sig_EN_stage = sig_EN_stage[order(sig_EN_stage[2]), ]
sig_EN_stage$peak_names = rownames(sig_EN_stage)
rownames(sig_EN_stage) = NULL # take out row names
sig_EN_stage = sig_EN_stage[c(3, 1, 2)]
DE_stages [["EN"]] <- sig_EN_stage

sig_BLC_stage <-
combined_df[combined_df$stagesBLC.adj.P.Val < 0.01 &
maxVals == 8, c(15, 16)]
sig_BLC_stage = sig_BLC_stage[order(sig_BLC_stage[2]), ]
sig_BLC_stage$peak_names = rownames(sig_BLC_stage)
rownames(sig_BLC_stage) = NULL # take out row names
sig_BLC_stage = sig_BLC_stage[c(3, 1, 2)]
DE_stages [["BLC"]] <- sig_BLC_stage


# saving results


DE_stages_logFC1 <- list()  # just interested in peaks with logFC>1

for (i in names(DE_stages)) {
DE_stages_logFC1[[i]] = DE_stages[[i]][which(DE_stages[[i]][[2]] > 1), ]
}


for (i in names(DE_stages_logFC1)) {
write.csv(
DE_stages_logFC1[i],
quote = F,
row.names = F,
file = paste(
"/Users/Marta/Documents/WTCHG/DPhil/Data/Results/ATAC-seq/DEA/peak",
currentDate,
"_sig_maxvals_",
i,
"_diff_peak_analysis_atac-seq_10CPM_trim_conservativeCounts_allQualitySamples.csv",
sep = ""
)
)

}



```
